requirements.txt
-i https://pypi.org/simple
argparse==1.4.0
cffi==1.16.0; python_version >= '3.8'
geographiclib==2.0; python_version >= '3.7'
geopy==2.4.1; python_version >= '3.7'
h3==3.7.7
numpy==2.0.0; python_version >= '3.9'
pycparser==2.22; python_version >= '3.8'
pyswisseph==2.10.3.2; python_version >= '3.5'
pytz==2024.1
setuptools==70.3.0; python_version >= '3.8'
tabulate==0.9.0; python_version >= '3.7'
timezonefinder==6.5.2; python_version >= '3.8' and python_version < '4'
